Fix curl-ev: migrating form to mime: 7
