+++ core/src/clients/http/plugin.cpp
@@ -12,16 +12,18 @@ PluginRequest::PluginRequest(RequestState& state) : state_(state) {}
 
 void PluginRequest::SetHeader(std::string_view name, std::string_view value) {
     state_.easy().add_header(
-        name, value, curl::easy::EmptyHeaderAction::kDoNotSend, curl::easy::DuplicateHeaderAction::kReplace
+        name, value, curl::detail::empty_header_action::kDoNotSend, curl::detail::duplicate_header_action::kReplace
     );
 }
 
 void PluginRequest::AddQueryParams(std::string_view params) {
     const auto& url = state_.easy().get_original_url();
     if (url.find('?') != std::string::npos) {
-        state_.easy().set_url(utils::StrCat(url, "&", params));
+        std::error_code ec;
+        state_.easy().set_url(utils::StrCat(url, "&", params), ec);
     } else {
-        state_.easy().set_url(utils::StrCat(url, "?", params));
+        std::error_code ec;
+        state_.easy().set_url(utils::StrCat(url, "?", params), ec);
     }
 }
