+++ core/src/clients/http/request.cpp
@@ -40,20 +40,24 @@ constexpr std::string_view kHeaderExpect = "Expect";
 
 std::string ToString(HttpMethod method) { return std::string{ToStringView(method)}; }
 
-curl::easy::http_version_t ToNative(HttpVersion version) {
+curl::detail::http_version_t ToNative(HttpVersion version) {
     switch (version) {
         case HttpVersion::kDefault:
-            return curl::easy::http_version_t::http_version_none;
+            return curl::detail::http_version_t::http_version_none;
         case HttpVersion::k10:
-            return curl::easy::http_version_t::http_version_1_0;
+            return curl::detail::http_version_t::http_version_1_0;
         case HttpVersion::k11:
-            return curl::easy::http_version_t::http_version_1_1;
+            return curl::detail::http_version_t::http_version_1_1;
         case HttpVersion::k2:
-            return curl::easy::http_version_t::http_version_2_0;
+            return curl::detail::http_version_t::http_version_2_0;
         case HttpVersion::k2Tls:
-            return curl::easy::http_version_t::http_version_2tls;
+            return curl::detail::http_version_t::http_version_2tls;
         case HttpVersion::k2PriorKnowledge:
-            return curl::easy::http_version_t::http_version_2_prior_knowledge;
+            return curl::detail::http_version_t::http_version_2_prior_knowledge;
+        case HttpVersion::k30:
+            return curl::detail::http_version_t::http_version_3_0;
+        case HttpVersion::k3Only:
+            return curl::detail::http_version_t::http_version_3_0_only;
     }
 
     UINVARIANT(false, "Unexpected HTTP version");
@@ -72,49 +76,49 @@ constexpr utils::TrivialBiMap kAuthTypeMap = [](auto selector) {
         .Case("any_safe", ProxyAuthType::kAnySafe);
 };
 
-curl::easy::httpauth_t HttpAuthTypeToNative(HttpAuthType value) {
+curl::detail::httpauth_t HttpAuthTypeToNative(HttpAuthType value) {
     switch (value) {
         case HttpAuthType::kBasic:
-            return curl::easy::auth_basic;
+            return curl::detail::httpauth_t::auth_basic;
         case HttpAuthType::kDigest:
-            return curl::easy::auth_digest;
+            return curl::detail::httpauth_t::auth_digest;
         case HttpAuthType::kDigestIE:
-            return curl::easy::auth_digest_ie;
+            return curl::detail::httpauth_t::auth_digest_ie;
         case HttpAuthType::kNegotiate:
-            return curl::easy::auth_negotiate;
+            return curl::detail::httpauth_t::auth_negotiate;
         case HttpAuthType::kNtlm:
-            return curl::easy::auth_ntlm;
+            return curl::detail::httpauth_t::auth_ntlm;
         case HttpAuthType::kNtlmWb:
-            return curl::easy::auth_ntlm_wb;
+            return curl::detail::httpauth_t::auth_ntlm_wb;
         case HttpAuthType::kAny:
-            return curl::easy::auth_any;
+            return curl::detail::httpauth_t::auth_any;
         case HttpAuthType::kAnySafe:
-            return curl::easy::auth_any_safe;
+            return curl::detail::httpauth_t::auth_any_safe;
     }
 
     UINVARIANT(false, "Unexpected http auth type");
 }
 
-curl::easy::proxyauth_t ProxyAuthTypeToNative(ProxyAuthType value) {
+curl::detail::proxyauth_t ProxyAuthTypeToNative(ProxyAuthType value) {
     switch (value) {
         case ProxyAuthType::kBasic:
-            return curl::easy::proxy_auth_basic;
+            return curl::detail::proxyauth_t::proxy_auth_basic;
         case ProxyAuthType::kDigest:
-            return curl::easy::proxy_auth_digest;
+            return curl::detail::proxyauth_t::proxy_auth_digest;
         case ProxyAuthType::kDigestIE:
-            return curl::easy::proxy_auth_digest_ie;
+            return curl::detail::proxyauth_t::proxy_auth_digest_ie;
         case ProxyAuthType::kBearer:
-            return curl::easy::proxy_auth_bearer;
+            return curl::detail::proxyauth_t::proxy_auth_bearer;
         case ProxyAuthType::kNegotiate:
-            return curl::easy::proxy_auth_negotiate;
+            return curl::detail::proxyauth_t::proxy_auth_negotiate;
         case ProxyAuthType::kNtlm:
-            return curl::easy::proxy_auth_ntlm;
+            return curl::detail::proxyauth_t::proxy_auth_ntlm;
         case ProxyAuthType::kNtlmWb:
-            return curl::easy::proxy_auth_ntlm_wb;
+            return curl::detail::proxyauth_t::proxy_auth_ntlm_wb;
         case ProxyAuthType::kAny:
-            return curl::easy::proxy_auth_any;
+            return curl::detail::proxyauth_t::proxy_auth_any;
         case ProxyAuthType::kAnySafe:
-            return curl::easy::proxy_auth_anysafe;
+            return curl::detail::proxyauth_t::proxy_auth_anysafe;
     }
 
     UINVARIANT(false, "Unexpected proxy auth type");
@@ -323,13 +327,13 @@ Request& Request::unix_socket_path(const std::string& path) & {
 Request Request::unix_socket_path(const std::string& path) && { return std::move(this->unix_socket_path(path)); }
 
 Request& Request::use_ipv4() & {
-    pimpl_->easy().set_ip_resolve(curl::easy::ip_resolve_v4);
+    pimpl_->easy().set_ip_resolve(curl::detail::ip_resolve_t::ip_resolve_v4);
     return *this;
 }
 Request Request::use_ipv4() && { return std::move(this->use_ipv4()); }
 
 Request& Request::use_ipv6() & {
-    pimpl_->easy().set_ip_resolve(curl::easy::ip_resolve_v6);
+    pimpl_->easy().set_ip_resolve(curl::detail::ip_resolve_t::ip_resolve_v6);
     return *this;
 }
 Request Request::use_ipv6() && { return std::move(this->use_ipv6()); }
@@ -341,18 +345,19 @@ Request& Request::connect_to(const ConnectTo& connect_to) & {
 Request Request::connect_to(const ConnectTo& connect_to) && { return std::move(this->connect_to(connect_to)); }
 
 Request& Request::data(std::string data) & {
-    if (!data.empty()) pimpl_->easy().add_header(kHeaderExpect, "", curl::easy::EmptyHeaderAction::kDoNotSend);
+    if (!data.empty()) pimpl_->easy().add_header(kHeaderExpect, "", curl::detail::empty_header_action::kDoNotSend);
     pimpl_->easy().set_post_fields(std::move(data));
     return *this;
 }
 Request Request::data(std::string data) && { return std::move(this->data(std::move(data))); }
 
-Request& Request::form(Form&& form) & {
-    pimpl_->easy().set_http_post(std::move(form).GetNative());
-    pimpl_->easy().add_header(kHeaderExpect, "", curl::easy::EmptyHeaderAction::kDoNotSend);
-    return *this;
-}
-Request Request::form(Form&& form) && { return std::move(this->form(std::move(form))); }
+//Request& Request::form(Form&& form) & {
+//!    pimpl_->easy().set_http_post(std::move(form).GetNative()); //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////!
+//    pimpl_->easy().add_header(kHeaderExpect, "", curl::detail::empty_header_action::kDoNotSend);
+//   return *this;
+//}
+
+//Request Request::form(Form&& form) && { return std::move(this->form(std::move(form))); }
 
 Request& Request::headers(const Headers& headers) & {
     SetHeaders(pimpl_->easy(), headers);
