+++ core/src/clients/http/request_state.cpp
@@ -114,8 +114,8 @@ void SetBaggageHeader(curl::easy& e) {
         e.add_header(
             USERVER_NAMESPACE::http::headers::kXBaggage,
             baggage->ToString(),
-            curl::easy::EmptyHeaderAction::kDoNotSend,
-            curl::easy::DuplicateHeaderAction::kReplace
+            curl::detail::empty_header_action::kDoNotSend,
+            curl::detail::duplicate_header_action::kReplace
         );
     }
 }
@@ -250,9 +250,7 @@ RequestState::RequestState(
 }
 
 RequestState::~RequestState() {
-    std::error_code ec;
-    easy().set_error_buffer(nullptr, ec);
-    UASSERT(!ec);
+    easy().set_error_buffer(nullptr);
 }
 
 void RequestState::follow_redirects(bool follow) {
@@ -275,8 +273,10 @@ void RequestState::ca(crypto::Certificate cert) {
     } else {
         // Legacy non-portable way, broken since 7.87.0
         ca_ = std::move(cert);
-        easy().set_ssl_ctx_function(&RequestState::on_certificate_request);
-        easy().set_ssl_ctx_data(this);
+        std::error_code ec;
+        easy().set_ssl_ctx_function(&RequestState::on_certificate_request, ec);
+        if (!ec)
+            easy().set_ssl_ctx_data(this, ec);
     }
 }
 
@@ -323,12 +323,14 @@ void RequestState::client_key_cert(crypto::PrivateKey pkey, crypto::Certificate 
         cert_id.resize(kCertIdLength, '=');
         easy().set_egd_socket(cert_id);
 
-        easy().set_ssl_ctx_function(&RequestState::on_certificate_request);
-        easy().set_ssl_ctx_data(this);
+        std::error_code ec;
+        easy().set_ssl_ctx_function(&RequestState::on_certificate_request, ec);
+        if (!ec)
+            easy().set_ssl_ctx_data(this, ec);
     }
 }
 
-void RequestState::http_version(curl::easy::http_version_t version) { easy().set_http_version(version); }
+void RequestState::http_version(curl::detail::http_version_t version) { easy().set_http_version(version); }
 
 void RequestState::set_timeout(long timeout_ms) {
     original_timeout_ = std::chrono::milliseconds{timeout_ms};
@@ -355,10 +357,10 @@ void RequestState::proxy(const std::string& value) {
     easy().set_proxy(value);
 }
 
-void RequestState::proxy_auth_type(curl::easy::proxyauth_t value) { easy().set_proxy_auth(value); }
+void RequestState::proxy_auth_type(curl::detail::proxyauth_t value) { easy().set_proxy_auth(value); }
 
 void RequestState::http_auth_type(
-    curl::easy::httpauth_t value,
+    curl::detail::httpauth_t value,
     bool auth_only,
     std::string_view user,
     std::string_view password
@@ -560,7 +562,7 @@ void RequestState::on_retry(std::shared_ptr<RequestState> holder, std::error_cod
         ++holder->retry_.current;
         holder->easy().mark_retry();
 
-        holder->retry_.timer.emplace(holder->easy().GetThreadControl());
+        holder->retry_.timer.emplace(holder->easy().get_thread_control());
 
         // call on_retry_timer on timer
         auto& holder_ref = *holder;
@@ -679,8 +681,10 @@ RequestState::async_perform_stream(const std::shared_ptr<Queue>& queue, utils::i
     auto& span = span_storage_->Get();
     span.AddTag("stream_api", 1);
 
-    easy().set_write_function(&RequestState::StreamWriteFunction);
-    easy().set_write_data(this);
+    std::error_code ec;
+    easy().set_write_function(&RequestState::StreamWriteFunction, ec);
+    if (!ec)
+        easy().set_write_data(this, ec);
     // Force no retries
     retry_.retries = 1;
 
@@ -774,7 +778,7 @@ void RequestState::UpdateTimeoutHeader() {
     easy().add_header(
         USERVER_NAMESPACE::http::headers::kXYaTaxiClientTimeoutMs,
         fmt::to_string(remote_timeout_.count()),
-        curl::easy::DuplicateHeaderAction::kReplace
+        curl::detail::duplicate_header_action::kReplace
     );
 }
