+++ core/src/curl-ev/form.cpp
@@ -24,14 +24,14 @@ form::~form() {
     }
 }
 
-/* void form::add_content(std::string_view key, std::string_view content) {
+void form::add_content(std::string_view key, std::string_view content) {
     std::error_code ec;
     add_content(key, content, ec);
     throw_error(ec, "add_content");
-} */
+}
 
-/* void form::add_content(std::string_view key, std::string_view content, std::error_code& ec) {
-    ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
+void form::add_content(std::string_view key, std::string_view content, std::error_code& ec) {
+/*     ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
         &post_,
         &last_,
         native::CURLFORM_COPYNAME,
@@ -43,22 +43,22 @@ form::~form() {
         native::CURLFORM_CONTENTSLENGTH,
         content.length(),
         native::CURLFORM_END
-    ))};
-} */
+    ))}; */
+}
 
-/* void form::add_content(std::string_view key, std::string_view content, const std::string& content_type) {
+void form::add_content(std::string_view key, std::string_view content, const std::string& content_type) {
     std::error_code ec;
     add_content(key, content, content_type, ec);
     throw_error(ec, "add_content");
-} */
+}
 
-/* void form::add_content(
+void form::add_content(
     std::string_view key,
     std::string_view content,
     const std::string& content_type,
     std::error_code& ec
 ) {
-    ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
+/*     ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
         &post_,
         &last_,
         native::CURLFORM_COPYNAME,
@@ -72,17 +72,17 @@ form::~form() {
         native::CURLFORM_CONTENTTYPE,
         content_type.c_str(),
         native::CURLFORM_END
-    ))};
-} */
+    ))}; */
+}
 
-/* void form::add_buffer(
+void form::add_buffer(
     const std::string& key,
     const std::string& file_name,
     const char* buffer,
     size_t buffer_len,
     std::error_code& ec
 ) {
-    ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
+/*     ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
         &post_,
         &last_,
         native::CURLFORM_COPYNAME,
@@ -96,10 +96,10 @@ form::~form() {
         native::CURLFORM_BUFFERLENGTH,
         buffer_len,
         native::CURLFORM_END
-    ))};
-} */
+    ))}; */
+}
 
-/* void form::add_buffer(
+void form::add_buffer(
     const std::string& key,
     const std::string& file_name,
     const std::shared_ptr<std::string>& buffer
@@ -107,9 +107,9 @@ form::~form() {
     std::error_code ec;
     add_buffer(key, file_name, buffer, ec);
     throw_error(ec, "add_buffer");
-} */
+}
 
-/* void form::add_buffer(
+void form::add_buffer(
     const std::string& key,
     const std::string& file_name,
     const std::shared_ptr<std::string>& buffer,
@@ -117,9 +117,9 @@ form::~form() {
 ) {
     buffers_.push_back(buffer);
     add_buffer(key, file_name, buffers_.back()->c_str(), buffers_.back()->size(), ec);
-} */
+}
 
-/* void form::add_buffer(
+void form::add_buffer(
     const std::string& key,
     const std::string& file_name,
     const std::shared_ptr<std::string>& buffer,
@@ -128,9 +128,9 @@ form::~form() {
     std::error_code ec;
     add_buffer(key, file_name, buffer, content_type, ec);
     throw_error(ec, "add_buffer");
-} */
+}
 
-/* void form::add_buffer(
+void form::add_buffer(
     const std::string& key,
     const std::string& file_name,
     const std::shared_ptr<std::string>& buffer,
@@ -138,7 +138,7 @@ form::~form() {
     std::error_code& ec
 ) {
     buffers_.push_back(buffer);
-    ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
+/*     ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
         &post_,
         &last_,
         native::CURLFORM_COPYNAME,
@@ -154,17 +154,17 @@ form::~form() {
         native::CURLFORM_CONTENTTYPE,
         content_type.c_str(),
         native::CURLFORM_END
-    ))};
-} */
+    ))}; */
+}
 
-/* void form::add_file(const std::string& key, const std::string& file_path) {
+void form::add_file(const std::string& key, const std::string& file_path) {
     std::error_code ec;
     add_file(key, file_path, ec);
     throw_error(ec, "add_file");
-} */
+}
 
-/* void form::add_file(const std::string& key, const std::string& file_path, std::error_code& ec) {
-    ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
+void form::add_file(const std::string& key, const std::string& file_path, std::error_code& ec) {
+/*     ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
         &post_,
         &last_,
         native::CURLFORM_COPYNAME,
@@ -174,22 +174,22 @@ form::~form() {
         native::CURLFORM_FILE,
         file_path.c_str(),
         native::CURLFORM_END
-    ))};
-} */
+    ))}; */
+}
 
-/* void form::add_file(const std::string& key, const std::string& file_path, const std::string& content_type) {
+void form::add_file(const std::string& key, const std::string& file_path, const std::string& content_type) {
     std::error_code ec;
     add_file(key, file_path, content_type, ec);
     throw_error(ec, "add_file");
-} */
+}
 
-/* void form::add_file(
+void form::add_file(
     const std::string& key,
     const std::string& file_path,
     const std::string& content_type,
     std::error_code& ec
 ) {
-    ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
+/*     ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
         &post_,
         &last_,
         native::CURLFORM_COPYNAME,
@@ -201,22 +201,22 @@ form::~form() {
         native::CURLFORM_CONTENTTYPE,
         content_type.c_str(),
         native::CURLFORM_END
-    ))};
-} */
+    ))}; */
+}
 
-/* void form::add_file_using_name(const std::string& key, const std::string& file_path, const std::string& file_name) {
+void form::add_file_using_name(const std::string& key, const std::string& file_path, const std::string& file_name) {
     std::error_code ec;
     add_file_using_name(key, file_path, file_name, ec);
     throw_error(ec, "add_file_using_name");
-} */
+}
 
-/* void form::add_file_using_name(
+void form::add_file_using_name(
     const std::string& key,
     const std::string& file_path,
     const std::string& file_name,
     std::error_code& ec
 ) {
-    ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
+/*     ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
         &post_,
         &last_,
         native::CURLFORM_COPYNAME,
@@ -228,10 +228,10 @@ form::~form() {
         native::CURLFORM_FILENAME,
         file_name.c_str(),
         native::CURLFORM_END
-    ))};
-} */
+    ))}; */
+}
 
-/* void form::add_file_using_name(
+void form::add_file_using_name(
     const std::string& key,
     const std::string& file_path,
     const std::string& file_name,
@@ -240,16 +240,16 @@ form::~form() {
     std::error_code ec;
     add_file_using_name(key, file_path, file_name, content_type, ec);
     throw_error(ec, "add_file_using_name");
-} */
+}
 
-/* void form::add_file_using_name(
+void form::add_file_using_name(
     const std::string& key,
     const std::string& file_path,
     const std::string& file_name,
     const std::string& content_type,
     std::error_code& ec
 ) {
-    ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
+/*     ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
         &post_,
         &last_,
         native::CURLFORM_COPYNAME,
@@ -263,17 +263,17 @@ form::~form() {
         native::CURLFORM_CONTENTTYPE,
         content_type.c_str(),
         native::CURLFORM_END
-    ))};
-} */
+    ))}; */
+}
 
-/* void form::add_file_content(const std::string& key, const std::string& file_path) {
+void form::add_file_content(const std::string& key, const std::string& file_path) {
     std::error_code ec;
     add_file_content(key, file_path, ec);
     throw_error(ec, "add_file_content");
-} */
+}
 
-/* void form::add_file_content(const std::string& key, const std::string& file_path, std::error_code& ec) {
-    ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
+void form::add_file_content(const std::string& key, const std::string& file_path, std::error_code& ec) {
+/*     ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
         &post_,
         &last_,
         native::CURLFORM_COPYNAME,
@@ -283,22 +283,22 @@ form::~form() {
         native::CURLFORM_FILECONTENT,
         file_path.c_str(),
         native::CURLFORM_END
-    ))};
-} */
+    ))}; */
+}
 
-/* void form::add_file_content(const std::string& key, const std::string& file_path, const std::string& content_type) {
+void form::add_file_content(const std::string& key, const std::string& file_path, const std::string& content_type) {
     std::error_code ec;
     add_file_content(key, file_path, content_type, ec);
     throw_error(ec, "add_file_content");
-} */
+}
 
-/* void form::add_file_content(
+void form::add_file_content(
     const std::string& key,
     const std::string& file_path,
     const std::string& content_type,
     std::error_code& ec
 ) {
-    ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
+/*     ec = std::error_code{static_cast<errc::FormErrorCode>(native::curl_formadd(
         &post_,
         &last_,
         native::CURLFORM_COPYNAME,
@@ -310,8 +310,8 @@ form::~form() {
         native::CURLFORM_CONTENTTYPE,
         content_type.c_str(),
         native::CURLFORM_END
-    ))};
-} */
+    ))}; */
+}
 
 }  // namespace curl
